pyproject.toml
[project]
name = "gmail-email-downloader"
version = "1.0.0"
description = "Download emails from Gmail labels using the Gmail API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.111.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
]

[project.scripts]
list-labels = "list_labels:main"
download-emails = "download_emails:main"

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 